Add support for Apple M1 in Clang configurations
This commit adds support for Apple M1 in Clang configurations by adding a new
global boolean flag `yaccs_APPLE_M1`.

This flag is set to `OFF` by default but can be overridden in the user config
file. This only affects Clang configurations at the moment and replace the
`-march` setting with `-mcpu=apple-m1`.